Assignment2/pyproject.toml
[project]
name = "assignment2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chronos-forecasting[torch]>=2.2.2",
    "matplotlib>=3.10.8",
    "optuna-dashboard>=0.20.0",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "tensorboard>=2.20.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
]
